requirements.txt
fastapi
uvicorn
pydantic
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
